app/requirements.txt
google-api-python-client
google-cloud-logging
google-cloud-bigquery
google-cloud-secret-manager
bs4
sendgrid
requests